+++ Src/Private/ConvertTo-Dot.ps1
@@ -42,10 +42,8 @@ function ConvertTo-Dot {
             Write-Debug -Message $($_.Exception.Message)
         }
         if ($Document) {
-            if ($Document) {
-                Write-Verbose -Message "Successfully converted Graphviz object to DOT format. Saved Path: $DestinationPath."
-                Get-ChildItem -Path $Document
-            }
+            Write-Verbose -Message "Successfully converted Graphviz object to DOT format. Saved Path: $DestinationPath."
+            Get-ChildItem -Path $Document
         }
     }
     end {}